requirements.txt
# MUST INSTALL INFORMER2020 MANUALLY AT THE END

# Core Data Processing
pandas
numpy==1.26.4
scipy

# Technical Indicators
ta-lib
pandas_ta

# Machine Learning & Deep Learning
scikit-learn
xgboost
lightgbm
pytorch-lightning
tqdm

# Bayesian Optimization & Probabilistic Models
pyro-ppl
scikit-optimize

# Graph Neural Networks (Market Microstructure)
networkx

# NLP & Sentiment Analysis
transformers
finbert

# Finance Data Sources
yfinance
alpha_vantage

# Reinforcement Learning for Trade Execution
gym

# Model Evaluation & Feature Selection
shap
ray[tune]
autogluon

# Logging, Experiment Tracking & Deployment
mlflow
tensorboard
onnx
onnxruntime

# Miscellaneous Utilities
matplotlib
seaborn
hydra-core
jupyter
ipywidgets